% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disj}
\alias{disj}
\title{disj}
\usage{
disj(strata, ncat)
}
\arguments{
\item{strata}{vector of integer}
}
\value{
vector
}
\description{
disjunctive
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
