% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disjMatrix}
\alias{disjMatrix}
\title{disjunctive Matrix}
\usage{
disjMatrix(strata)
}
\arguments{
\item{A}{matrix of integer that contains categorical variable in column.}
}
\value{
A matrix of indicators.
}
\description{
This function transform a categorical matrix into a matrix of indicators variables.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
