% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastcubeArma}
\alias{fastcubeArma}
\title{fastcubeArma}
\usage{
fastcubeArma(X, Xcat, pik)
}
\arguments{
\item{X}{matrix}

\item{Xcat}{matrix}

\item{pik}{matrix}
}
\value{
sample
}
\description{
fastcubeArma
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
