% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findB.R
\name{findB}
\alias{findB}
\title{Title}
\usage{
findB(X, Xcat)
}
\arguments{
\item{Xcat}{}
}
\value{

}
\description{
Title
}
\examples{
rm(list = ls())

library(sampling)

N <- 1000

Xcat <-data.frame(cat1 = rep(1:40,each = N/40),
                    cat2 = rep(1:50,each = N/50),
                    cat3 = rep(1:500,times = 2))
                    #cat3 = rep(1:100,each = N/100))

pik <- inclusionprobastrata(Xcat[,1],rep(1,40))
p <- 30
X <- matrix(rnorm(N*p),ncol = 30)
}
