% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findBarma}
\alias{findBarma}
\title{findBarma}
\usage{
findBarma(X, Xcat)
}
\arguments{
\item{X}{a Matrix}

\item{Xcat}{a Matrix}
}
\value{
a matrix
}
\description{
findB
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
