% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flightphase_arma}
\alias{flightphase_arma}
\title{title}
\usage{
flightphase_arma(X, pik, redux)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{pik}{vector of inclusion probabilities}

\item{EPS}{tolerance}
}
\value{
a sample
}
\description{
title
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
