% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flightphaseModif}
\alias{flightphaseModif}
\title{flightphaseModif}
\usage{
flightphaseModif(prob, Xbal)
}
\arguments{
\item{prob}{vector of inclusion probabilities of size N.}

\item{Xbal}{Matrix of auxiliary variables of dimension N x p}
}
\value{
a sample with at most p value not update to 0 or 1.
}
\description{
flightphaseModif
}
