% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isEye2}
\alias{isEye2}
\title{is Identiy matrix}
\usage{
isEye2(M)
}
\arguments{
\item{M}{matrix}
}
\value{
bool
}
\description{
is Identiy matrix
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
