% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncat}
\alias{ncat}
\title{number of category}
\usage{
ncat(Xcat)
}
\arguments{
\item{Xcat}{A matrix of integer that contains categorical variable in column.}
}
\value{
A row vector that contains the number of category in each column.
}
\description{
This function returns the number of factor in each column of a categorical matrix.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
