% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{onestep2}
\alias{onestep2}
\title{title}
\usage{
onestep2(B, pik, EPS = 1e-07)
}
\arguments{
\item{B}{matrix of auxiliary variables.}

\item{pik}{vector of inclusion probabilities}

\item{EPS}{tolerance}
}
\value{
updated pik
}
\description{
description
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
