% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osffphase2}
\alias{osffphase2}
\title{title}
\usage{
osffphase2(prob, Bm)
}
\arguments{
\item{prob}{inclusion probabilities}

\item{Bm}{matrix of auxiliary variables}
}
\value{
a vector
}
\description{
description
}
\details{
details
}
\examples{

}
\seealso{
func
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
