% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rea}
\alias{rea}
\title{title}
\usage{
rea(index, pik, done, howlong)
}
\arguments{
\item{index}{}
}
\value{
index updated index
}
\description{
title
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
