% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samplen}
\alias{samplen}
\title{title}
\usage{
samplen(N, n)
}
\arguments{
\item{x}{x}
}
\value{
a vector
}
\description{
description Rosetta code combn
}
\details{
details
}
\examples{

}
\seealso{
func
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
