% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{systematicDesign}
\alias{systematicDesign}
\title{systematic desgin}
\usage{
systematicDesign(pik)
}
\arguments{
\item{pik}{vector of inclusion probabilities..}
}
\value{
A matrix of size at most N x N with elements equal to 0 or 1. The value 1 indicates that the uni is selected while the value 0 is for non-chosen unit.
}
\description{
Find all possible systematic sample from a vector of inclusion probabilities \code{pik}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
