% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disj}
\alias{disj}
\title{disjunctive vector}
\usage{
disj(strata)
}
\arguments{
\item{strata}{A vector of integer that represents the category.}
}
\value{
A matrix of indicators.
}
\description{
This function transform a categorical variable into a matrix of indicators.
}
\examples{
  strata <- rep(c(1,2,3),each = 4)
  disj(strata)

}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
