% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disjMatrix}
\alias{disjMatrix}
\title{disjMatrix}
\usage{
disjMatrix(strata)
}
\arguments{
\item{starta}{a Matrix}
}
\value{
a matrix
}
\description{
disj on each column
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
