% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findB.R
\name{findB}
\alias{findB}
\title{find the sub-matrix B in the flight phase}
\usage{
findB(X, Xcat)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{Xcat}{matrix of categorical variables.}
}
\value{
matrix B
}
\description{
This function is used in the \code{\link{ffphase}} function to find the smallest matrix B.
}
\details{
The function find the smallest matrix B such that is contains only one additional row.
 It consecutively add the right number of row depending of the number of categroies that is added.
}
\examples{
rm(list = ls())

library(sampling)

N <- 1000

Xcat <-data.frame(cat1 = rep(1:40,each = N/40),
                    cat2 = rep(1:50,each = N/50),
                    cat3 = rep(1:500,times = 2))
                    #cat3 = rep(1:100,each = N/100))

pik <- inclusionprobastrata(Xcat[,1],rep(1,40))
p <- 30
X <- matrix(rnorm(N*p),ncol = 30)
}
