% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncat}
\alias{ncat}
\title{ncat}
\usage{
ncat(Xcat)
}
\arguments{
\item{Xcat}{Matrix}
}
\value{
vector
}
\description{
number of cat in each column
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
